Swap names of quicksort and timsort generic and templated files
