+++ numpy/_core/meson.build
@@ -1230,9 +1230,9 @@ src_multiarray = multiarray_gen_headers + [
   'src/multiarray/temp_elide.c',
   'src/multiarray/usertypes.c',
   'src/multiarray/vdot.c',
-  'src/npysort/quicksort.cpp',
+  'src/npysort/quicksort_generic.cpp',
   'src/npysort/mergesort.cpp',
-  'src/npysort/timsort.cpp',
+  'src/npysort/timsort_generic.cpp',
   'src/npysort/heapsort.cpp',
   'src/npysort/radixsort.cpp',
   'src/npysort/npysort_methods.cpp',
